Implemented `FILL_BYTES` compressor

+++ include/compression-context.h
@@ -6,7 +6,7 @@ typedef struct {
 	Header *restrict small;
 	Header *restrict large;
 	const u8 *restrict uncompressed;
-	size_t uncompressedSize;
+	u16 uncompressedSize;
 } CompressionContext;
 void initCompressionContext(CompressionContext *restrict context, const u8 *restrict uncompressed, size_t uncompressedSize);
 void freeCompressionContext(CompressionContext *restrict context);
